refactor: move FormalArbiterTester
